+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-# 0.11.0 2012/12/2
+# 0.11.1 2012/12/3
 
 Feature
 - add `etherchain`, `etherchain-tx` skills and `gaspriceoracle` skill
 
 Enhancement
-- [beaconchain][etherchaintx][bitquerytx] simplify the regex
+- [beaconchain][etherchaintx][bitquerytx][gastracker][gasstation][gasnow][gaspriceoracle][address][etherchain][etherscan][etherscantx] simplify the regex
 
 Fix
 - [xdai-tx][bsc-tx][blockchair] fix the skill and simplify the regex